Added additional cleaning out.
For some reason some of the image files fail to be deleted as a consequence
of the git stash command due to permission reasons, which is odd (they should
be run with the same user ID) and I don't know how to correct. We're hoping
that the rm -f gets rid of them. Perhaps we just shouldn't bother returning
to master at the end of deployment.