turned on eval and modify data import.
VDS had eval = FALSE for all the chunks so that they did not run.
this would cause the vignette to build even if it had errors.

I have also modified data import in both vignettes to show what
the user would likely do, but import from the data directory
behind the scenes.