Various small fixes and edits.
These are the result of a review up to _Genetic diversity (osberved and expected heterozygosity)_. I was careful not to change the meaning of anything that was said.

Some of the changes are corrections of mistakes in the text. For example, the text said that the subsetting would keep 20 SNPs, whereas the code shown kept only 11. I decided to go with what the code said.