Removes vignette on basic population genetic statistics.
This is in essence a repeat of 85951c0bb8524f34cfce9e6332e2b04f36d13bed,
this time on the master branch, because master had changes to the vignette
(however small), creating a merge conflict. Repeating the deletion here
prior to merge should have the advantage that we keep a commit that we can
simply revert on a development branch to reinstate the vignette from its
latest stage.