Migrates CircleCI configuration to version 2
In practice this isn't a migration so much as a rewrite, as the CircleCI
configuration changed quite drastically from version 1 to version 2.

As a result, rewriting the config file alone is in fact far from sufficient,
and making this work precipitates a number of changes in how the process is
run. More specifically, originally the CircleCI configuration replicated
the notion that we are on a machine that can build and run Docker containers.
This mirrored the building locally situation. In version 2, this is still
possible (by using "machine: true"), but more natural and the long term path
is that we *are* in "the" (or a) container.

We could still by choice of primary container and using multiple containers
stay in the paradigm of building our container on the fly and then running it
repeatedly. But that seemed to get more complex than needed, and it works
more naturally in version 2 to simply do everything in the target container.

In terms of the target container, we were using the one built on the fly
as essentially a content-less wrapper around the hlapp/rpopgen container
from Docker Hub, so this is what we are changing to.

Moves CircleCI config to the correct name